Move polyfill route wrapping logic to serve.py.
