+++ tools/wptrunner/wptrunner/wptcommandline.py
@@ -241,7 +241,7 @@ scheme host and port.""")
     config_group.add_argument("--font-dir", action="store", type=abs_path, dest="font_dir",
                               help="Path to local font installation directory", default=None)
     config_group.add_argument("--polyfill", action="store", dest="polyfill", default=None,
-                              help="URL to polyfill to inject")
+                              help="Path to polyfill to inject")
     config_group.add_argument("--headless", action="store_true",
                               help="Run browser in headless mode", default=None)
     config_group.add_argument("--no-headless", action="store_false", dest="headless",
