+++ tools/wptserve/wptserve/wptserve.py
@@ -19,8 +19,6 @@ def parse_args():
     parser.add_argument("--host", "-H", dest="host", action="store",
                         type=str, default="127.0.0.1",
                         help="Host to run server on")
-    parser.add_argument("--polyfill", action="store", dest="polyfill", default=None,
-                        help="URL of polyfill to inject")
     return parser.parse_args()
 
 
@@ -28,8 +26,7 @@ def main():
     args = parse_args()
     httpd = WebTestHttpd(host=args.host, port=args.port,
                          use_ssl=False, certificate=None,
-                         doc_root=args.document_root,
-                         polyfill=args.polyfill)
+                         doc_root=args.document_root)
     httpd.start()
 
 if __name__ == "__main__":
